backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
supabase==2.4.6
anthropic==0.28.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
httpx==0.27.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
bcrypt==4.1.3
